refactor(studio): extract scaling helpers and sql-helpers from jobs.ts
